requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# LLM Providers
openai>=1.0.0
google-generativeai>=0.3.0
ollama>=0.1.0

# Embeddings & Similarity
sentence-transformers>=2.2.0
rouge-score>=0.1.2

# Evaluation Metrics
scipy>=1.10.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Data Loading (HuggingFace datasets)
datasets>=2.14.0
huggingface-hub>=0.16.0
